requirements.txt
grpcio==1.42.0
torch==2.2.0
torchaudio==2.2.0
numpy==1.24.3
soundfile==0.12.1
transformers==4.37.2
